--- src/Application/Import/ImportActivityLaps/ImportActivityLaps.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Application\Import\ImportActivityLaps;
-
-use App\Infrastructure\CQRS\Command\DomainCommand;
-use Symfony\Component\Console\Output\OutputInterface;
-
-final readonly class ImportActivityLaps extends DomainCommand
-{
-    public function __construct(
-        private OutputInterface $output,
-    ) {
-    }
-
-    public function getOutput(): OutputInterface
-    {
-        return $this->output;
-    }
-}
