--- src/Application/Import/ImportActivitySplits/ImportActivitySplits.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Application\Import\ImportActivitySplits;
-
-use App\Infrastructure\CQRS\Command\DomainCommand;
-use Symfony\Component\Console\Output\OutputInterface;
-
-final readonly class ImportActivitySplits extends DomainCommand
-{
-    public function __construct(
-        private OutputInterface $output,
-    ) {
-    }
-
-    public function getOutput(): OutputInterface
-    {
-        return $this->output;
-    }
-}
